pyproject.toml
[tool.poetry]
name = "youtube-video-downloader"
version = "0.0.1"
description = ""
authors = ["Pedro Augusto Barbosa Aparecido <pedr.augustobarbosa.aparecido@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11,<3.13"
google-api-python-client = "^2.108.0"
google-auth-oauthlib = "^1.1.0"
google-auth-httplib2 = "^0.1.1"
python-youtube = "^0.9.2"
loguru = "^0.7.2"
appdata = "^2.2.0"
userpaths = "^0.1.3"
pandas = "^2.1.3"
pyxlsb = "^1.0.10"
pytube = "^15.0.0"
requests = "^2.31.0"
clean-text = "^0.6.0"
unidecode = "^1.3.7"
tqdm = "^4.66.1"
pydub = "^0.25.1"
moviepy = "^1.0.3"
ffmpeg = "^1.4"
ffmpeg-python = "^0.2.0"
more-itertools = "^10.1.0"
aioitertools = "^0.11.0"
pillow = "^10.1.0"
plyer = "^2.1.0"
winotify = "^1.1.0"
psgtray = "^1.0.2"
pywebview = "^4.4.1"


[tool.poetry.group.examples.dependencies]
python-dotenv = "^1.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
